chore(web): rename nested SiCanvas directory to canvas
This commit solely renames the nested SiCanvas directory to canvas. No
files nore directories were moved.
