+++ app/web/src/observable/selection.ts
@@ -1,7 +1,7 @@
 import * as Rx from "rxjs";
 
-import { SceneManager } from "@/organisms/SiCanvas/SiCanvas/scene_manager";
-import { Node } from "@/organisms/SiCanvas/SiCanvas/obj/node";
+import { SceneManager } from "@/organisms/SiCanvas/canvas/scene_manager";
+import { Node } from "@/organisms/SiCanvas/canvas/obj/node";
 
 // These shouldn't be global, interaction manager should own them
