+++ app/web/src/organisms/PanelAttribute.vue
@@ -157,7 +157,7 @@
 </template>
 
 <script setup lang="ts">
-import { Node } from "@/organisms/SiCanvas/SiCanvas/obj/node";
+import { Node } from "@/organisms/SiCanvas/canvas/obj/node";
 import * as Rx from "rxjs";
 import _ from "lodash";
