+++ app/web/src/organisms/SiCanvas/canvas/interaction_manager/panning.ts
