+++ app/web/src/organisms/SiCanvas/canvas/scene_manager.ts
